# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Larave",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "vyacheslavkarmazin@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/token/login": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Token",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/token/refresh": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Token",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthController::refresh",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Return posts",
                "description": "Get posts",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\PostsController::getAll",
                "responses": {
                    "200": {
                        "description": "Array of post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "post id",
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "title": {
                                                "description": "Title of post",
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "date": {
                                                "description": "Create date",
                                                "type": "string",
                                                "example": "2017-01-01",
                                                "nullable": false
                                            },
                                            "text": {
                                                "description": "text of post",
                                                "type": "string",
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Return post",
                "description": "Create post",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\PostsController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of post",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "text": {
                                        "description": "text of post",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of posts"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/post/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Return post",
                "description": "Get post",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\PostsController::getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "post id",
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "title": {
                                            "description": "Title of post",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "date": {
                                            "description": "Create date",
                                            "type": "string",
                                            "example": "2017-01-01",
                                            "nullable": false
                                        },
                                        "text": {
                                            "description": "text of post",
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Return post",
                "description": "Create post",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\PostsController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of post",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "text": {
                                        "description": "text of post",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "post id",
                                            "type": "integer",
                                            "nullable": true
                                        },
                                        "title": {
                                            "description": "Title of post",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "date": {
                                            "description": "Create date",
                                            "type": "string",
                                            "example": "2017-01-01",
                                            "nullable": false
                                        },
                                        "text": {
                                            "description": "text of post",
                                            "type": "string",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Value: Bearer {jwt}",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}